requirements.txt
pypdf>=5.4.0
python-pptx>=1.0.2
python-docx>=1.1.2
reportlab>=4.2.5
